+++ invoice-service/src/main/java/com/team28/booking/invoice/repository/DiscountUsageProjection.java
@@ -1,14 +1,15 @@
 package com.team28.booking.invoice.repository;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface DiscountUsageProjection {
     Long getDiscountId();
     String getCode();
     String getDiscountType();
-    Double getDiscountValue();
+    BigDecimal getDiscountValue();
     Integer getTimesUsed();
-    Double getTotalDiscountGiven();
+    BigDecimal getTotalDiscountGiven();
     Boolean getActive();
     LocalDateTime getExpiryDate();
 }
